# internal/api/docs/swagger.json
{
    "components": {
        "schemas": {
...
            "immerle.ErrorResponse": {
                "properties": {
                    "error": {
                        "example": "already_initialized",
                        "type": "string"
                    },
                    "ok": {
                        "example": false,
                        "type": "boolean"
                    }
                },
                "type": "object"
            },
...
            "immerle.JamParticipantDTO": {
                "properties": {
                    "joinedAt": {
                        "example": "2026-06-15T09:00:00Z",
                        "type": "string"
                    },
                    "sessionId": {
                        "type": "string"
                    },
                    "userId": {
                        "type": "string"
                    },
                    "username": {
                        "example": "bob",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "immerle.JamResponse": {
                "properties": {
                    "ok": {
                        "example": true,
                        "type": "boolean"
                    },
                    "participants": {
                        "items": {
                            "$ref": "#/components/schemas/immerle.JamParticipantDTO"
                        },
                        "type": "array",
                        "uniqueItems": false
                    },
                    "session": {
                        "$ref": "#/components/schemas/immerle.JamSessionDTO"
                    }
                },
                "type": "object"
            },
            "immerle.JamSessionDTO": {
                "properties": {
                    "currentTrackId": {
                        "type": "string"
                    },
                    "hostId": {
                        "type": "string"
                    },
                    "id": {
                        "type": "string"
                    },
                    "name": {
                        "example": "Friday Night",
                        "type": "string"
                    },
                    "position": {
                        "example": 42000,
                        "type": "integer"
                    },
                    "state": {
                        "example": "playing",
                        "type": "string"
                    },
                    "trackIds": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array",
                        "uniqueItems": false
                    }
                },
                "type": "object"
            },
...
            "immerle.OKResponse": {
                "properties": {
                    "ok": {
                        "example": true,
                        "type": "boolean"
                    }
                },
                "type": "object"
            },
...
        }
    },
    "info": {
        "contact": {
            "name": "immerle"
        },
        "description": "Native immerle endpoints (first-run setup, capability\ndiscovery, friends, activity feed, collaborative playlists and\nsynchronized Jam sessions) that complement the Subsonic /\nOpenSubsonic API served under /rest/.\n\nAuthenticated endpoints reuse Subsonic credentials passed as\nquery parameters: u (username) plus either p (password) or the\ntoken pair t (md5(password+salt)) and s (salt), and c (client\nname). Setup and capability endpoints are unauthenticated.",
        "license": {
            "name": "See repository"
        },
        "title": "immerle extension API",
        "version": "1.0.0"
    },
...
    "paths": {
...
        "/jam/create": {
            "post": {
                "description": "Starts a synchronized listening session hosted by the caller.",
                "parameters": [
                    {
                        "description": "Subsonic username",
                        "in": "query",
                        "name": "u",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Subsonic password (or t+s token auth)",
                        "in": "query",
                        "name": "p",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Client name",
                        "in": "query",
                        "name": "c",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Session name",
                        "in": "query",
                        "name": "name",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Comma-separated track ids",
                        "in": "query",
                        "name": "trackIds",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/immerle.JamResponse"
                                }
                            }
                        },
                        "description": "OK"
                    }
                },
                "summary": "Create a Jam session",
                "tags": [
                    "jam"
                ]
            }
        },
        "/jam/events": {
            "get": {
                "description": "Server-Sent Events stream. Emits the current state immediately, then a \"state\"/\"participants\"/\"closed\" event on every change.",
                "parameters": [
                    {
                        "description": "Subsonic username",
                        "in": "query",
                        "name": "u",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Subsonic password (or t+s token auth)",
                        "in": "query",
                        "name": "p",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Client name",
                        "in": "query",
                        "name": "c",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Jam session id",
                        "in": "query",
                        "name": "sessionId",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "text/event-stream": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        },
                        "description": "SSE stream"
                    },
                    "404": {
                        "content": {
                            "text/event-stream": {
                                "schema": {
                                    "$ref": "#/components/schemas/immerle.ErrorResponse"
                                }
                            }
                        },
                        "description": "Not Found"
                    }
                },
                "summary": "Stream Jam events (SSE)",
                "tags": [
                    "jam"
                ]
            }
        },
        "/jam/join": {
            "post": {
                "parameters": [
                    {
                        "description": "Subsonic username",
                        "in": "query",
                        "name": "u",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Subsonic password (or t+s token auth)",
                        "in": "query",
                        "name": "p",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Client name",
                        "in": "query",
                        "name": "c",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Jam session id",
                        "in": "query",
                        "name": "sessionId",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/immerle.JamResponse"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "404": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/immerle.ErrorResponse"
                                }
                            }
                        },
                        "description": "Not Found"
                    }
                },
                "summary": "Join a Jam session",
                "tags": [
                    "jam"
                ]
            }
        },
        "/jam/leave": {
            "post": {
                "parameters": [
                    {
                        "description": "Subsonic username",
                        "in": "query",
                        "name": "u",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Subsonic password (or t+s token auth)",
                        "in": "query",
                        "name": "p",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Client name",
                        "in": "query",
                        "name": "c",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Jam session id",
                        "in": "query",
                        "name": "sessionId",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/immerle.OKResponse"
                                }
                            }
                        },
                        "description": "OK"
                    }
                },
                "summary": "Leave a Jam session",
                "tags": [
                    "jam"
                ]
            }
        },
        "/jam/state": {
            "get": {
                "parameters": [
                    {
                        "description": "Subsonic username",
                        "in": "query",
                        "name": "u",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Subsonic password (or t+s token auth)",
                        "in": "query",
                        "name": "p",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Client name",
                        "in": "query",
                        "name": "c",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Jam session id",
                        "in": "query",
                        "name": "sessionId",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/immerle.JamResponse"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "404": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/immerle.ErrorResponse"
                                }
                            }
                        },
                        "description": "Not Found"
                    }
                },
                "summary": "Get Jam session state",
                "tags": [
                    "jam"
                ]
            }
        },
        "/jam/update": {
            "post": {
                "description": "Host-only. Updates the shared track/position/state and broadcasts it to participants over SSE.",
                "parameters": [
                    {
                        "description": "Subsonic username",
                        "in": "query",
                        "name": "u",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Subsonic password (or t+s token auth)",
                        "in": "query",
                        "name": "p",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Client name",
                        "in": "query",
                        "name": "c",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Jam session id",
                        "in": "query",
                        "name": "sessionId",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Current track id",
                        "in": "query",
                        "name": "currentTrackId",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Playback position in ms",
                        "in": "query",
                        "name": "position",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "description": "playing or paused",
                        "in": "query",
                        "name": "state",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Comma-separated track ids",
                        "in": "query",
                        "name": "trackIds",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/immerle.JamResponse"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "403": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/immerle.ErrorResponse"
                                }
                            }
                        },
                        "description": "Forbidden"
                    },
                    "404": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/immerle.ErrorResponse"
                                }
                            }
                        },
                        "description": "Not Found"
                    }
                },
                "summary": "Update Jam playback (host only)",
                "tags": [
                    "jam"
                ]
            }
        },
...
    },
    "openapi": "3.1.0",
...
}